Rework event system
Now <obj> and <instance> classes can register a ON-EVENT method which
will be called when a event is triggered. For toggling to work there
is a <toggleable> class now, that defines a ON-CLICK callback and a
TOGGLED? method which can be used to check if a obj is toggled. This
comes with the caveat that one hast to inherit from <toggleable>
first. This also requires persistent instance's though, hence the
rewrite of some PROCESS methods with GET-INSTANCES and
UPDATE-SLOTS. ON-EVENT's return value is used to determine if UPDATE
should be called, this is useful to update the data before printing.